implement base view
