+++ django_app/core/base_views.py
@@ -1,12 +1,17 @@
 import datetime
+from typing import Any
 
 from django import forms
 from django.conf import settings
+from django.http import HttpRequest, HttpResponse
 from django.shortcuts import redirect
+from django.template.loader import render_to_string
 from django.urls import reverse
 from django.utils import timezone
 from django.utils.http import url_has_allowed_host_and_scheme
-from django.views.generic import FormView, TemplateView
+from django.utils.safestring import mark_safe
+from django.views.generic import DetailView, FormView, TemplateView
+from playwright.sync_api import sync_playwright
 from report_a_suspected_breach.utils import get_dirty_form_data
 
 
@@ -98,3 +103,33 @@ class BaseFormView(FormView):
 
 class BaseTemplateView(TemplateView):
     pass
+
+
+class BaseDownloadPDFView(DetailView):
+    template_name = "core/base_download_pdf.html"
+    header = "Report a suspected breach"
+
+    def get(self, request: HttpRequest, *args: object, **kwargs: object) -> HttpResponse:
+        context_data = self.get_context_data(**kwargs)
+        filename = f"report-{context_data['reference']}.pdf"
+        pdf_data = None
+        template_string = render_to_string(self.template_name, context=context_data)
+
+        with sync_playwright() as playwright:
+            browser = playwright.chromium.launch(headless=True)
+            page = browser.new_page()
+            page.set_content(mark_safe(template_string))
+            page.wait_for_function("document.fonts.ready.then(fonts => fonts.status === 'loaded')")
+            pdf_data = page.pdf(format="A4")
+            browser.close()
+
+        response = HttpResponse(pdf_data, content_type="application/pdf")
+        response["Content-Disposition"] = f"inline; filename={filename}"
+        return response
+
+    def get_context_data(self, *args: object, **kwargs: object) -> dict[str, Any]:
+        context = super().get_context_data(**kwargs)
+        context["header"] = self.header
+        reference = self.request.GET.get("reference", "")
+        context["reference"] = reference
+        return context
