+++ django_app/core/templates/core/base_download_pdf.html
@@ -1,0 +1,15 @@
+{% extends "core/base.html" %}
+{% load get_css_rules_as_string %}
+{% block cookie_banner %}
+{% endblock cookie_banner %}
+{% block header_content %}
+{% endblock header_content %}
+{% block skip_link %}{% endblock skip_link %}
+{% block phase_banner %}{% endblock phase_banner %}
+{% block back_button %}{% endblock back_button %}
+<script>
+    {% get_css_rules_as_string "core/stylesheets/font-awesome-6.5.2.css" %}
+    {% get_css_rules_as_string "core/stylesheets/govuk-frontend-5.3.0.min.css" %}
+    {% get_css_rules_as_string "core/stylesheets/common.css" %}
+</script>
+{% block footer %}{% endblock footer %}
