+++ django_app/report_a_suspected_breach/urls.py
@@ -3,7 +3,6 @@ from report_a_suspected_breach.views import (
     generic,
     views_business,
     views_documents_and_details,
-    views_download,
     views_end,
     views_sanctions_and_goods,
     views_start,
@@ -124,7 +123,7 @@ views_end_urls = [
     path("check-your-answers", views_end.CheckYourAnswersView.as_view(), name="check_your_answers"),
     path("declaration", views_end.DeclarationView.as_view(), name="declaration"),
     path("submission-complete", views_end.CompleteView.as_view(), name="complete"),
-    path("download_report/", views_download.DownloadPDFView.as_view(), name="download_report"),
+    path("download_report/", views_end.DownloadPDFView.as_view(), name="download_report"),
 ]
 
 urlpatterns = (
