+++ django_app/view_a_suspected_breach/views.py
@@ -1,19 +1,17 @@
 import logging
 from typing import Any
 
+from core.base_views import BaseDownloadPDFView
 from core.sites import require_view_a_breach
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.auth.models import User
 from django.db.models import QuerySet
 from django.http import HttpRequest, HttpResponse, HttpResponseRedirect
 from django.shortcuts import get_object_or_404, reverse
-from django.template.loader import render_to_string
 from django.urls import reverse_lazy
 from django.utils.decorators import method_decorator
-from django.utils.safestring import mark_safe
 from django.views.generic import DetailView, ListView, RedirectView, TemplateView
 from feedback.models import FeedbackItem
-from playwright.sync_api import sync_playwright
 from report_a_suspected_breach.models import Breach
 from utils.breach_report import get_breach_context_data
 
@@ -122,29 +120,13 @@ class ViewFeedbackView(LoginRequiredMixin, ActiveUserRequiredMixin, DetailView):
 
 
 @method_decorator(require_view_a_breach(), name="dispatch")
-class DownloadPDFView(LoginRequiredMixin, ActiveUserRequiredMixin, DetailView):
+class DownloadPDFView(LoginRequiredMixin, ActiveUserRequiredMixin, BaseDownloadPDFView):
     template_name = "view_a_suspected_breach/viewer_pdf.html"
-
-    def get(self, request: HttpRequest, *args: object, **kwargs: object) -> HttpResponse:
-        context_data = self.get_context_data(**kwargs)
-        filename = f"report-{context_data['reference']}.pdf"
-        pdf_data = None
-        template_string = render_to_string(self.template_name, context=context_data)
-
-        with sync_playwright() as playwright:
-            browser = playwright.chromium.launch(headless=True)
-            page = browser.new_page()
-            page.set_content(mark_safe(template_string))
-            page.wait_for_function("document.fonts.ready.then(fonts => fonts.status === 'loaded')")
-            pdf_data = page.pdf(format="A4")
-            browser.close()
-
-        response = HttpResponse(pdf_data, content_type="application/pdf")
-        response["Content-Disposition"] = f"inline; filename={filename}"
-        return response
+    header = "Report a suspected breach of trade sanctions: report submitted"
 
     def get_context_data(self, *args: object, **kwargs: object) -> dict[str, Any]:
-        self.object = get_object_or_404(Breach, reference=self.request.GET.get("reference", ""))
+        reference = self.request.GET.get("reference")
+        self.object = get_object_or_404(Breach, reference=reference)
         context = super().get_context_data(**kwargs)
         breach_context_data = get_breach_context_data(self.object)
         context.update(breach_context_data)
